pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "carapace"
version = "0.1.0"
description = "PR and issue triage orchestration with similarity clustering and canonical ranking"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
  { name = "Carapace Contributors" }
]
dependencies = [
  "pydantic>=2.6",
  "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4"
]

[project.scripts]
carapace = "carapace.cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["carapace*"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
